+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
  - Cooldown to most commands
  - Autodelete of balance commands
  - Autodelete of withdrawal commands
+ - Discord logging (via logging module)
 
 ## Changed
  - Withdrawal command to have a specific on reaction loop instead of a general loop